Fix Product and Cart, create and update bug
For any field on api side that is nullable-true and or has a default value, pass `|| undefined` along the data passed for create and update. As in { field: field || undefined }

Consider one of the methods

read = (productId: number) => {
    console.log({ type: typeof productId });
    const req = new AxiosBaseRequest(`${this.path}/${productId}`);
    this.makeRequest(req);
};

run: `npm start -- rp 1`

output
```
{ type: 'string' }
{
  id: 1,
  createdAt: '2022-01-02T12:01:42.878Z',
  updatedAt: '2022-01-02T12:07:58.269Z',
  name: 'Chicken soup with broccoli',
  description: 'This is a native chicken with broccoli soup from new your',
  price: '£6.99',
  count: 3
}
```

The type is string even thought the type of productId is number or expects a number type of value
Consider the method below that is called in the above snippet

```
// Read a Product by productId
App.command("readProduct <productId>")
    .aliases(["rp", "product"])
    .description("Read a Product by productId")
    .action((productId: number) => productWrapper.read(productId));
```

in the action method, productId is of type number so why is it a string?

Cast numeric type value to number using Number (it appears that even though the type of a variable is number, an input from the terminal returns a string into a numberic type of value)

Product

create: npm start -- cp "Chicken soup with broccoli" "" 12 5
{ success: true, message: 'Created successfully' }

read all: npm start -- lp
[
  {
    id: 1,
    createdAt: '2022-01-02T12:01:42.878Z',
    updatedAt: '2022-01-02T12:01:42.878Z',
    name: 'Chicken soup with broccoli',
    description: 'Sample product',
    price: '£12.00',
    count: 5
  }
]

update description: npm start -- up 1 "" "This is a native chicken with broccoli soup from new your" 0 0
{ success: true, message: 'Updated successfully' }

update price and count: npm start -- up 1 "" "" 6.99 3
{ success: true, message: 'Updated successfully' }

read all: npm start -- lp
[
  {
    id: 1,
    createdAt: '2022-01-02T12:01:42.878Z',
    updatedAt: '2022-01-02T12:07:58.269Z',
    name: 'Chicken soup with broccoli',
    description: 'This is a native chicken with broccoli soup from new your',
    price: '£6.99',
    count: 3
  }
]